deployment_automation_agent/requirements.txt
docker
python-dotenv
pika
